database/requirements.txt
alembic==1.13.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pandas==2.1.4
python-dotenv==1.0.0 